Use base handler for sync account job
